backend/requirements.txt
Django
psycopg2-binary
django-oauth-toolkit
djangorestframework
oauthlib
requests